Added network structure and framework for NN
The neural network was starting to get hard to think about. Although
others have build NN in "11 lines of python" it was actually for a NN
_without_ any hidden layers. The fact that I have hidden layers means
that mine is inheritly more complex.

To battle my confusion designing the NN I decided to create a class that
controls the nodes and one that controls the synapses connecting the
nodes. The nodes themselves are stored in 4 dictionaries: the input
dict, hidden dict, bias dict and output dict.