Initialised "Simple Neurons" notebook
I have found that starting on fully-fledged neural networks was a
terrible idea. They are incredibly complex, and I can't wrap my head
around the big idea. As such, I've decided to start on a simple in/out
NN with *no* hidden layers.

To understand what I mean, I've included a quick sketch of what I plan
on designing below:

A
 \
  \
   O----C
  /
 /
B

Where A and B are input neurons and C is the output neuron.